Detect native session loading
If a session is loaded via the commandline, it is not normally detected
by sessionist, and the name needs to be re-specified to save it.
This change allows sessionist to detect a natively loaded session from
vim, and use the specified path as the session path.

To facilitate this, a new variable has been introduced called
`g:current_session_path`, which specifies the full path of the current
session. `g:current_session` is still used, and each are derived from
the other as necessary.

A number of new functions have also been added to dedupe functionality